fix(zcncore): added transaction_query_base
